requirements.txt
streamlit
pandas
plotly
pydeck
Pillow
numpy